Commit 2.14*: Fix styles for buttons components in DialogContact for mobile view and 'window.confirm' logic.

+++ part2/phonebook/src/components/ContactDialog.jsx
@@ -38,7 +38,10 @@ const ContactDialog = forwardRef(
     }; // closes the dialog and resets the state
 
     const handleDelete = () => {
-      window.confirm(`Are you sure you want to delete ${contactDetails.name}?`);
+      const isDeletable = confirm(
+        "Are you sure you want to delete this contact?"
+      );
+      if (!isDeletable) return;
       deleteContact(contactDetails.id)
         .then((data) => {
           handleClosure(undefined);
